requirements.txt
crowdstrike-falconpy
google-api-python-client
